requirements.txt
fastapi 
uvicorn 
paddleocr 
pillow 
opencv-python-headless 
numpy
python-multipart
pdf2image